문자열 계산기: 리팩토링
