+++ src/main/kotlin/step2/tokenize/DefaultTokenizer.kt
@@ -1,0 +1,3 @@
+package step2.tokenize
+
+class DefaultTokenizer : Tokenize
