+++ src/main/kotlin/step2/tokenize/Tokenize.kt
@@ -1,0 +1,7 @@
+package step2.tokenize
+
+interface Tokenize {
+    operator fun invoke(equation: String, delimiter: String): List<String> {
+        return equation.split(delimiter)
+    }
+}
